ml/app/requirements.txt
fastapi
uvicorn
Pillow
diffusers
segment-anything
torch
torchvision
torchaudio